carify and facilitate tf remote state creation
- adds context to change into directory with `GithubConfiguration.yaml`
  file (I missed that the first time)

- adds a `tf-state.sh` script that users can run

It seems odd that in a post talking about Terraform, we're asking users
to use Cloudformation to deploy something :)  I know of several
Terraform shops that have almost no knowledge or experience with CFN so
this might feel very strange.  I'd be happy to port this over to
Terraform and provide instructions if you're interested.
Just let me know.